+++ site/pages/our-work/case-study/bbc/index.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import WhatToReadNext from '../what-to-read-next';
+import WhatToReadNext from '../../what-to-read-next';
 import ChecklistContactUs from '../../../../slices/checklist-contact-us-slice';
 
 import CaseStudyFrame from '../../../../components/case-study-frame';
