+++ site/pages/our-work/case-study/fortnum-and-mason-digital-transformation/index.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import Card from '../../../../components/card';
 import ListBox from '../../../../components/list-box';
-import WhatToReadNext from '../what-to-read-next';
+import WhatToReadNext from '../../what-to-read-next';
 import ChecklistContactUs from '../../../../slices/checklist-contact-us-slice';
 import Social from '../../../../components/social';
 import Picture from '../../../../components/picture';
