+++ site/pages/our-work/case-study/sky/index.js
@@ -12,7 +12,7 @@ import Card from '../../../../components/card';
 import ListBox from '../../../../components/list-box';
 import Quote from '../../../../components/quote';
 
-import WhatToReadNext from '../what-to-read-next';
+import WhatToReadNext from '../../what-to-read-next';
 import ChecklistContactUs from '../../../../slices/checklist-contact-us-slice';
 import ContactBox from '../../../../components/contact-box';
 import ScrollTracker from '../../../../components/scroll-tracker';
